+++ cmd/tsserver/main.go
@@ -1,12 +1,15 @@
 package tsserver
 
 import (
+	"log/slog"
+
 	"github.com/direktiv/direktiv/pkg/tsengine"
 )
 
 func RunApplication() {
 	srv, err := tsengine.NewServer()
 	if err != nil {
+		slog.Error("failed to start the ts-engine server", "error", err)
 		panic(err)
 	}
