+++ pkg/service/manager.go
@@ -12,6 +12,7 @@ import (
 	"github.com/direktiv/direktiv/pkg/core"
 	"github.com/direktiv/direktiv/pkg/reconcile"
 	"github.com/direktiv/direktiv/pkg/tracing"
+	"github.com/pkg/errors"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"knative.dev/serving/pkg/client/clientset/versioned"
@@ -142,7 +143,7 @@ func (m *manager) Run(circuit *core.Circuit) error {
 		errs := m.runCycle()
 		m.lock.Unlock()
 		for _, err := range errs {
-			slog.Error("run cycle", "err", err, "cycleFails", cycleFails)
+			slog.Error("run cycle", "err", err)
 		}
 
 		// Evaluate errors rate.
@@ -150,6 +151,9 @@ func (m *manager) Run(circuit *core.Circuit) error {
 			cycleFails++
 		} else {
 			cycleFails = 0
+		}
+		if cycleFails > 5 {
+			return errors.New("too many cycle fails")
 		}
 
 		time.Sleep(cycleTime)
